refactor: dedup of components
