+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Card/CollectiblePaper.tsx
@@ -7,6 +7,7 @@ export interface CollectiblePaperProps extends withClasses<'root'> {
 export function CollectiblePaper(props: CollectiblePaperProps) {
     const { children } = props
     const classes = props.classes
+
     return (
         <Card className={classes?.root} elevation={0}>
             {children}
