+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Card/tabs/DetailTab.tsx
@@ -2,10 +2,10 @@ import { useMemo } from 'react'
 import type { AsyncState } from 'react-use/lib/useAsyncFn'
 import { LoadingBase, makeStyles } from '@masknet/theme'
 import type { Web3Helper } from '@masknet/plugin-infra/web3'
-import { Context } from '../hooks/useContext.js'
 import { CollectibleCard } from '../CollectibleCard.js'
-import { DetailsCard } from '../../CardDialog/DetailsCard.js'
-import { PropertiesCard } from '../../CardDialog/PropertiesCard.js'
+import { DetailsCard } from '../../Shared/DetailsCard.js'
+import { PropertiesCard } from '../../Shared/PropertiesCard.js'
+import { Context } from '../../Context/index.js'
 
 const useStyles = makeStyles()((theme) => ({
     body: {
@@ -26,7 +26,7 @@ export interface DetailTabProps {
 
 export function DetailTab(props: DetailTabProps) {
     const { asset } = props
-    const { provider, rarity } = Context.useContainer()
+    const { sourceType, rarity } = Context.useContainer()
     const { classes } = useStyles()
 
     return useMemo(() => {
@@ -42,7 +42,7 @@ export function DetailTab(props: DetailTabProps) {
             <CollectibleCard>
                 <div className={classes.body}>
                     <div className={classes.info}>
-                        <DetailsCard sourceType={provider} asset={asset.value} />
+                        <DetailsCard sourceType={sourceType} asset={asset.value} />
                     </div>
                     <PropertiesCard rank={rarity.value?.rank} asset={asset.value} />
                 </div>
