+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Shared/ActivityCard.tsx
