+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Shared/CollectibleProviderIcon.tsx
