+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Shared/DescriptionCard.tsx
