+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Shared/DetailsCard.tsx
