+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Shared/PriceCard.tsx
