+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Shared/PropertiesCard.tsx
