+++ packages/mask/src/plugins/Collectible/SNSAdaptor/Shared/Rank.tsx
